Clarify LangChain integration module names
