+++ packages/langchain/src/index.ts
@@ -1,7 +1,7 @@
-export { createLangChainStreamResponse } from "./langchain-stream-response";
+export { createLangChainStreamResponse } from "./request-handler";
 export type {
   CreateLangChainStreamResponseOptions,
   LangChainInputMessage,
-} from "./langchain-stream-response";
-export { streamOpenUI } from "./stream-openui";
-export type { StreamOpenUIOptions } from "./stream-openui";
+} from "./request-handler";
+export { streamOpenUI } from "./stream-relay";
+export type { StreamOpenUIOptions } from "./stream-relay";
